# Conflicts:
#	apps/health_campaign_field_worker_app/lib/blocs/hcm_attendance_bloc.dart
#	apps/health_campaign_field_worker_app/lib/data/local_store/sql_store/sql_store.g.dart
#	apps/health_campaign_field_worker_app/lib/models/data_model.mapper.g.dart
#	apps/health_campaign_field_worker_app/lib/pages/beneficiary_registration/individual_details.dart
#	packages/attendance_management/lib/blocs/attendance_individual_bloc.freezed.dart
#	packages/attendance_management/lib/models/attendance_register.mapper.g.dart
#	packages/attendance_management/lib/models/attendee.mapper.g.dart
#	packages/attendance_management/lib/pages/manage_attendance.dart
#	packages/attendance_management/lib/pages/session_select.dart

+++ packages/attendance_management/lib/models/attendee.dart
@@ -32,7 +32,7 @@ class AttendeeModel with AttendeeModelMappable {
   final bool? nonRecoverableError;
   final String? name;
   final int? rowVersion;
-  final int? status;
+  final double? status;
   final String? identifierID;
   final String? individualNumber;
   final AttendeeAdditionalFields? additionalFields;
